// src/payload-types.ts
export interface Category {
  id: number;
  name: string;
  slug: string;
  color?: string | null;
  parent?: (number | null) | Category;
  subcategories?: {
    docs?: (number | Category)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}